scaler-bus-api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
exceptiongroup==1.3.1
fastapi==0.135.2
firebase_admin==7.3.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
